Stopped using python docker module
